Use a map/filter approach for generating occurrences
Initially this used a two step approach: generate values from the interval and
then run the adjustments on those interval elements. The second part however
does not always work, especially with a rule such as "every day in january,
daily" as it would not know to wrap around to the following year at the end of
january.

Instead, this now uses a map/filter approach that still generates single
interval elements, but will then create a chain of operations on this element
depending on what specific rules are specified. This properly handles the above
mentioned case as it would instead filter out elements that don't match the
month january and it would keep trying until either the next year starts or a
limit is reached.

This obviously comes with the problem of generating (possibly) endlessly as well
 as unnecessarily. To combat this, though yet not implement, filters can provide
 "suggestions" as to when the next entry might be, thus allowing the interval to
 skip ahead.

Additionally, this implements better testing by parsing the given examples in
the RFC and making sure those values match. For the 'forever' rules, this uses a
 different library to which the results will be compared. For the time being,
this only handles date occurrences from the RFC; time occurrences are not yet
included.